Better modularized pdf_parser
